1:修改停止监测时出现闪退问题 2：修改平均速度和行驶距离的数据类型 3：修改行驶距离的计算方法

+++ app/src/main/java/com/wxthxy/driving/database/LocationModel.java
@@ -34,13 +34,13 @@ public class LocationModel extends BaseModel implements Serializable{
      * km
      */
     @Column
-    public long totalmileage;
+    public double totalmileage;
     /**
      * 平均速度
      * km/h
      */
     @Column
-    public long averageVelocity;
+    public double averageVelocity;
 
     /**
      * 车辆掉头
